# Optimize set membership tests against lists.