--- a/policy/parser/yml/parser.go
+++ b/policy/parser/yml/parser.go
@@ -141,6 +141,7 @@ func (p *parser) parse(node *yaml.Node, ref objRef) {
 		p.parsePrimitive(node, ref)
 	}
 	ref.encodeStyle(getEncodeStyle(node.Style))
+	ref.finalize()
 }
 
 func (p *parser) parsePrimitive(node *yaml.Node, ref objRef) {