Since policy list values are constructed specially, held resident in memory, and usually don't contain many entries it is reasonable to build a value set into a go `map` which can be used for accelerating the `Contains` call used by the `in` operator.

Here are some of the before and after for the existing policy set:
```
benchmark                                                                       old ns/op     new ns/op     delta
BenchmarkEnforcer/restricted_destinations_valid_location-8                      4327          2812          -35.01%
BenchmarkEnforcer/restricted_destinations_valid_location_ir_label-8             4293          2965          -30.93%
BenchmarkEnforcer/dependent_ranges_behavior-8                                   8186          7015          -14.30%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     2713          2370          -12.64%
BenchmarkEnforcer/binauthz_package_violations-8                                 14708         13221         -10.11%

benchmark                                                                       old allocs     new allocs     delta
BenchmarkEnforcer/restricted_destinations_valid_location-8                      18             8              -55.56%
BenchmarkEnforcer/restricted_destinations_valid_location_ir_label-8             19             12             -36.84%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     14             12             -14.29%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  29             27             -6.90%
BenchmarkEnforcer/binauthz_package_violations-8                                 83             78             -6.02%

benchmark                                                                       old bytes     new bytes     delta
BenchmarkEnforcer/restricted_destinations_valid_location-8                      336           176           -47.62%
BenchmarkEnforcer/restricted_destinations_valid_location_ir_label-8             360           248           -31.11%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     280           248           -11.43%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  640           608           -5.00%
BenchmarkEnforcer/binauthz_package_violations-8                                 2490          2410          -3.21%
```